+++ packages/cli/src/generator_templates/ts/emails/previews/ResetPassword.tsx
@@ -9,7 +9,7 @@ export function resetPassword() {
           We’ve received your request to change your password. Use the link
           below to set up a new password for your account. This link is only
           usable once! If you need to, you can reinitiate the password process
-          again <a href="#">here</a>.
+          again <a href="https://www.mailing.run">here</a>.
         </>
       }
       ctaText="Reset Password"